4.21: only creator of a blog can delete the blog

+++ controllers/blogs.js
@@ -35,6 +35,16 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET_KEY)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid'})
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if (user.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'no rights to remove blog'})
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
